backend/requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
greenlet==3.1.1
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.10.0
redis==5.2.1
SQLAlchemy==2.0.38
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
